非DB化の修正対応

+++ install.php
@@ -48,12 +48,12 @@ if($_POST["install"]){
     );
 
     if(empty($_POST["domain_name"])){
-        $_POST["domain_name"] = "localhost";
+        $_POST["domain_name"] = $_SERVER["SERVER_NAME"];
     }
     if(!is_dir($baseDir."/_configure/")){
         mkdir($baseDir."/_configure/");
     }
-    if(($fp = fopen($baseDir."/_configure/configure_".$_POST["domain_name"].".php", "w+")) !== FALSE){
+    if(($fp = fopen($baseDir."/_configure/configure.php", "w+")) !== FALSE){
         $content = file_get_contents("http://config.vizualizer.jp/index.php", false, stream_context_create($context));
         echo nl2br(htmlspecialchars($content));
         fwrite($fp, $content);
@@ -72,32 +72,47 @@ if($_POST["install"]){
     Vizualizer_Bootstrap::register(60, "UserAgent");
     Vizualizer_Bootstrap::startup();
 
-    $connection = Vizualizer_Database_Factory::begin("default");
-    try{
-        $connection->query("DROP TABLE admin_company_operators");
-        $connection->query(file_get_contents($baseDir."/sql/01_admin_company_operators.sql"));
-        $connection->query("DROP TABLE admin_companys");
-        $connection->query(file_get_contents($baseDir."/sql/01_admin_companys.sql"));
-        $connection->query("DROP TABLE admin_roles");
-        $connection->query(file_get_contents($baseDir."/sql/01_admin_roles.sql"));
-        $connection->query(file_get_contents($baseDir."/sql/02_admin_company_operators.sql"));
-        $connection->query("INSERT INTO `admin_companys` (`company_id`, `company_name`, `display_flg`, `create_time`, `update_time`) VALUES (1, '".$_POST["company_name"]."', 1, NOW(), NOW())");
-        $connection->query("INSERT INTO `admin_roles` (`role_id`, `role_code`, `role_name`, `create_time`, `update_time`) VALUES (1, 'administrator', '管理者', NOW(), NOW()), (2, 'user', '利用者', NOW(), NOW())");
-        $connection->query("INSERT INTO `admin_company_operators` (`operator_id`, `company_id`, `role_id`, `login_id`, `password`, `operator_name`, `url`, `email`, `display_flg`, `create_time`, `update_time`) VALUES (1, 1, 2, '".$_POST["login_id"]."', SHA1('".$_POST["login_id"].":".$_POST["password"]."'), '".$_POST["operator_name"]."', 'link', '".$_POST["site_email"]."', 1, NOW(), NOW())");
-        Vizualizer_Database_Factory::commit($connection);
-    }catch(Vizualizer_Exception_Database $e){
-        Vizualizer_Database_Factory::rollback($connection);
-    }
+    // データベースを初期化する。
+    if($_POST["database_type"] == "mysql"){
+        $connection = Vizualizer_Database_Factory::begin("default");
+        try{
+            $connection->query("DROP TABLE admin_company_operators");
+            $connection->query(file_get_contents($baseDir."/sql/01_admin_company_operators.sql"));
+            $connection->query("DROP TABLE admin_companys");
+            $connection->query(file_get_contents($baseDir."/sql/01_admin_companys.sql"));
+            $connection->query("DROP TABLE admin_roles");
+            $connection->query(file_get_contents($baseDir."/sql/01_admin_roles.sql"));
+            $connection->query(file_get_contents($baseDir."/sql/02_admin_company_operators.sql"));
+            $connection->query("INSERT INTO `admin_companys` (`company_id`, `company_name`, `display_flg`, `create_time`, `update_time`) VALUES (1, '".$_POST["company_name"]."', 1, NOW(), NOW())");
+            $connection->query("INSERT INTO `admin_roles` (`role_id`, `role_code`, `role_name`, `create_time`, `update_time`) VALUES (1, 'administrator', '管理者', NOW(), NOW()), (2, 'user', '利用者', NOW(), NOW())");
+            $connection->query("INSERT INTO `admin_company_operators` (`operator_id`, `company_id`, `role_id`, `login_id`, `password`, `operator_name`, `url`, `email`, `display_flg`, `create_time`, `update_time`) VALUES (1, 1, 2, '".$_POST["login_id"]."', SHA1('".$_POST["login_id"].":".$_POST["password"]."'), '".$_POST["operator_name"]."', 'link', '".$_POST["site_email"]."', 1, NOW(), NOW())");
+            Vizualizer_Database_Factory::commit($connection);
+        }catch(Vizualizer_Exception_Database $e){
+            Vizualizer_Database_Factory::rollback($connection);
+        }
 
-    $connection = Vizualizer_Database_Factory::begin("shortage");
-    try{
-        $connection->query("DROP TABLE shortage_click_logs");
-        $connection->query(file_get_contents($baseDir."/sql/01_shortage_click_logs.sql"));
-        $connection->query("DROP TABLE shortage_urls");
-        $connection->query(file_get_contents($baseDir."/sql/01_shortage_urls.sql"));
-        Vizualizer_Database_Factory::commit($connection);
-    }catch(Vizualizer_Exception_Database $e){
-        Vizualizer_Database_Factory::rollback($connection);
+        $connection = Vizualizer_Database_Factory::begin("shortage");
+        try{
+            $connection->query("DROP TABLE shortage_click_logs");
+            $connection->query(file_get_contents($baseDir."/sql/01_shortage_click_logs.sql"));
+            $connection->query("DROP TABLE shortage_urls");
+            $connection->query(file_get_contents($baseDir."/sql/01_shortage_urls.sql"));
+            Vizualizer_Database_Factory::commit($connection);
+        }catch(Vizualizer_Exception_Database $e){
+            Vizualizer_Database_Factory::rollback($connection);
+        }
+    }elseif($_POST["database_type"] == "sqlite"){
+        copy(VIZUALIZER_SITE_ROOT."/db/shorturl.db", VIZUALIZER_SITE_ROOT."/database.db");
+        $connection = Vizualizer_Database_Factory::begin("default");
+        try{
+            $connection->query("INSERT INTO `admin_companys` (`company_id`, `company_name`, `display_flg`, `create_time`, `update_time`) VALUES (1, '".$_POST["company_name"]."', 1, '".date("Y-m-d H:i:s")."', '".date("Y-m-d H:i:s")."')");
+            $connection->query("INSERT INTO `admin_roles` (`role_id`, `role_code`, `role_name`, `create_time`, `update_time`) VALUES (1, 'administrator', '管理者', '".date("Y-m-d H:i:s")."', '".date("Y-m-d H:i:s")."')");
+            $connection->query("INSERT INTO `admin_roles` (`role_id`, `role_code`, `role_name`, `create_time`, `update_time`) VALUES (2, 'user', '利用者', '".date("Y-m-d H:i:s")."', '".date("Y-m-d H:i:s")."')");
+            $connection->query("INSERT INTO `admin_company_operators` (`operator_id`, `company_id`, `role_id`, `login_id`, `password`, `operator_name`, `url`, `email`, `display_flg`, `create_time`, `update_time`) VALUES (1, 1, 2, '".$_POST["login_id"]."', '".sha1($_POST["login_id"].":".$_POST["password"])."', '".$_POST["operator_name"]."', 'link', '".$_POST["site_email"]."', 1, '".date("Y-m-d H:i:s")."', '".date("Y-m-d H:i:s")."')");
+            Vizualizer_Database_Factory::commit($connection);
+        }catch(Vizualizer_Exception_Database $e){
+            Vizualizer_Database_Factory::rollback($connection);
+        }
     }
 
     echo "インストールが完了しました。";
@@ -125,15 +140,25 @@ if($_POST["install"]){
             <div class="center span5">
                 <div class="well spa12 login-box">
                     <form action="install.php" method="post">
+                        <input type="hidden" name="company_name" value="テストグループ" />
+                        <input type="hidden" name="operator_name" value="管理アカウント" />
+                        <input type="hidden" name="site_email" value="info@dot-hacks.com" />
+                        <input type="hidden" name="database[default][file]" value="database.db" />
                         <fieldset>
                             <div><label>
-                            利用ドメイン
-                            <input type="text" name="domain_name" value="" />
+                            ログインID
+                            <input type="text" name="login_id" value="" />
                             </label></div>
                             <div><label>
-                            メールアドレス
-                            <input type="text" name="site_email" value="" />
+                            パスワード
+                            <input type="text" name="password" value="" />
                             </label></div>
+                            <div><label>
+                            <input type="radio" name="database_type" value="sqlite" checked onclick="document.getElementById('database_setting').style.display = 'none'" />データベースを利用しない
+                            </label><label>
+                            <input type="radio" name="database_type" value="mysql" onclick="document.getElementById('database_setting').style.display = 'block'" />データベースを利用する
+                            </label></div>
+                            <div id="database_setting" style="display: none">
                             <div><label>
                             DB接続先
                             <input type="text" name="database[default][host]" value="" />
@@ -154,22 +179,7 @@ if($_POST["install"]){
                             短縮URL用DB名（オプション）
                             <input type="text" name="database[shortage][name]" value="" />
                             </label></div>
-                            <div><label>
-                            管理法人名
-                            <input type="text" name="company_name" value="" />
-                            </label></div>
-                            <div><label>
-                            管理者名
-                            <input type="text" name="operator_name" value="" />
-                            </label></div>
-                            <div><label>
-                            ログインID
-                            <input type="text" name="login_id" value="" />
-                            </label></div>
-                            <div><label>
-                            パスワード
-                            <input type="text" name="password" value="" />
-                            </label></div>
+                            </div>
                             <p class="center span5">
                                 <button type="submit" name="install" value="install">インストール</button>
                             </p>
